refactor: split transcript search worker lifecycles
